requirements.txt
numpy
pillow~=8.4.0
opencv-python~=4.0
imageio-ffmpeg